// program/clients/typescript/src/generated/pdas/payment.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { getAddressEncoder, getProgramDerivedAddress, getU32Encoder, getUtf8Encoder, type Address, type ProgramDerivedAddress } from '@solana/kit';


export type PaymentSeeds = {
                  
      merchantOperatorConfig: Address;
                
      buyer: Address;
                
      mint: Address;
                
      orderId: number;
      }

export async function findPaymentPda(
      seeds: PaymentSeeds,
    config: { programAddress?: Address | undefined } = {},
): Promise<ProgramDerivedAddress> {
  const { programAddress = 'commkU28d52cwo2Ma3Marxz4Qr9REtfJtuUfqnDnbhT' as Address<'commkU28d52cwo2Ma3Marxz4Qr9REtfJtuUfqnDnbhT'> } = config;
  return await getProgramDerivedAddress({ programAddress, seeds: [
                  getUtf8Encoder().encode("payment"),
                        getAddressEncoder().encode(seeds.merchantOperatorConfig),
                        getAddressEncoder().encode(seeds.buyer),
                        getAddressEncoder().encode(seeds.mint),
                        getU32Encoder().encode(seeds.orderId),
            ]});
}